:wrench:

+++ packages/skia/cpp/api/JsiSkFontStyle.h
@@ -38,15 +38,15 @@ public:
     } else {
       auto weightProp = object.getProperty(runtime, "weight");
       auto weight = static_cast<int>(weightProp.isUndefined()
-                                         ? SkFontStyle::Weight::kNormal_Weight
+                                         ? static_cast<double>(SkFontStyle::Weight::kNormal_Weight)
                                          : weightProp.asNumber());
       auto widthProp = object.getProperty(runtime, "width");
       auto width = static_cast<int>(widthProp.isUndefined()
-                                        ? SkFontStyle::Width::kNormal_Width
+                                        ? static_cast<double>(SkFontStyle::Width::kNormal_Width)
                                         : widthProp.asNumber());
       auto slantProp = object.getProperty(runtime, "slant");
       auto slant = static_cast<SkFontStyle::Slant>(
-          slantProp.isUndefined() ? SkFontStyle::Slant::kUpright_Slant
+          slantProp.isUndefined() ? static_cast<double>(SkFontStyle::Slant::kUpright_Slant)
                                   : slantProp.asNumber());
       SkFontStyle style(weight, width, slant);
       return std::make_shared<SkFontStyle>(style);
